requirements.txt
streamlit
pandas
numpy
plotly
supabase
python-dotenv
openai
